Move router to outter scope
Router is not a RTP entity, but a section of the QuickAPI's API.
